+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Allow adding maps to image collection tilesets (#3447)
 * Added file system actions to the tile context menu (#3448)
+* Fixed possible crash in Custom Types Editor (#3465)
 * Fixed ability to reset nested string and file properties (#3409)
 * Fixed case where object labels could become visible for hidden layer (#3442)
 * AutoMapping: Fixed applying of rule probability (#3425)